Dependency injection fixup
Inject an actual instance of the desired CliHelper subclass into Cli's
run() method, rather than a class. Applications are now responsible for
creating the helper, not the Cli. The CliHelper's __init__ signature is
now an empty argument list which allows applications to create
subclasses with whatever argument list is appropriate AND create the
helper whenever it's convenient. Previously, the helper's constructor
took a "cli" instance which mean helpers had to be constructed after the
Cli. This was accomplished by having the Cli perform the CliHelper
instance creation.

In the future, a database instance will be injected as well, so this
change ensures a measure of consistency. A design that asks clients
to inject a class of one thing, but an instance of another would be
confusing.