Cli single-responsibility cleanup: remove arg parsing.
Argument parsing has been moved into a simple module-level helper
function so that the arguments can be more accessible to applications
and they can create modify the parser as need be. This paves the
way for injecting the db object, different subclasses of which
may take different kinds of arguments. Doing the arg parsing
inside Cli facilitated writing very short applications, but
was not flexible enough.