Add special Cli subclass for File-based CLIs
Add a subclass of Cli that ensures that errors from File construction
are caught. Having it centralized in a subclass ensures that the
try-except needed by all applications doesn't need to be repeated.
Needing this was unexpected (oops, bad password causes nasty uncaught
exception)and seems a little odd. We'll see how it ages.

Also, add some class and function comment headers.