Move the password file path construction.
Move this out of the database file class and into two other places: up
a level into the interactive CLI class and into the "dumppw"
program. The reason is that *most* cases are served well by the
existing design, which allows the user to simply specify the name of a
password collection, but there's one case (the "dumppw" program) where
it's most convenient to specify the actual path to the data file.